fixes slider bugs
