+++ js/form.js
@@ -1,5 +1,7 @@
 import { body } from './big-picture-modal-open.js';
 import { isEscapeKey } from './util.js';
+import { sliderBackground, image, scale, onScaleChange } from './slider.js';
+// import { image } from './scale.js';
 
 const uploadButton = document.querySelector('#upload-file');
 const imageOverlay = document.querySelector('.img-upload__overlay');
@@ -34,6 +36,8 @@ function validateTags (value) {
 const showImageOverlay = function () {
   imageOverlay.classList.remove('hidden');
   body.classList.add('modal-open');
+  sliderBackground.classList.add('hidden');
+  scale.addEventListener('click', onScaleChange);
 };
 
 document.addEventListener('keydown', (evt) => {
@@ -41,6 +45,10 @@ document.addEventListener('keydown', (evt) => {
     evt.preventDefault();
     imageOverlay.classList.add('hidden');
     body.classList.remove('modal-open');
+    uploadButton.removeEventListener('change', showImageOverlay);
+    form.reset();
+    pristine.reset();
+    scale.removeEventListener('click', onScaleChange);
   }
 });
 pristine.addValidator(hashTagField, validateTags, ERROR_TEXT);
@@ -53,24 +61,15 @@ form.addEventListener('submit', (evt) => {
 });
 
 const closeImageOverlay = function () {
+  image.className = 'effects__preview--none';
   imageOverlay.classList.add('hidden');
   body.classList.remove('modal-open');
   uploadButton.removeEventListener('change', showImageOverlay);
   form.reset();
-  // resetScale();
-  // resetEffects();
   pristine.reset();
+  scale.removeEventListener('click', onScaleChange);
+  image.style.transform = 'scale(1)';
 };
-
-// function stopEvtPropagation (field1) {
-//   field1.addEventListener('keydown', (evt) => {
-//     if (isEscapeKey(evt)) {
-//       evt.stopPropagation();
-//     }
-//   });
-// }
-
-// stopEvtPropagation(hashTagField);
 
 commentField.addEventListener('keydown', (evt) => {
   if (isEscapeKey(evt)) {
@@ -85,7 +84,11 @@ hashTagField.addEventListener('keydown', (evt) => {
 });
 
 closeButton.addEventListener('click', closeImageOverlay);
-uploadButton.addEventListener('input', showImageOverlay);
 
-export {uploadButton, closeButton};
+const openImageForm = function () {
+  uploadButton.addEventListener('input', showImageOverlay);
+};
+
+
+export { openImageForm };
