+++ js/main.js
@@ -1,20 +1,13 @@
 import { addsPicturesTemplates } from './random-user-pictures.js';
 import { opensBigPictureModal } from './big-picture-modal-open.js';
 import { closeBigPicture, closeBigPictureByKey } from './big-picure-modal-close.js';
-import { uploadButton, closeButton } from './form.js';
-import { image } from './scale.js';
-import { slider } from './slider.js';
+import { openImageForm } from './form.js';
+import { onEffectsChange } from './slider.js';
 
 addsPicturesTemplates();
 opensBigPictureModal();
 closeBigPicture();
 closeBigPictureByKey();
-
-// const bigPictureContainer = document.querySelector('.big-picture');
-
-// const smallPicture = document.querySelector('.picture__img');
-// smallPicture.addEventListener('click', () => {
-//   bigPictureContainer.classList.remove('hidden');
-// });
-
+openImageForm();
+onEffectsChange();
